Tarea2/requirements.txt
confluent-kafka
grpcio
grpcio-tools
elasticsearch

